moving map

+++ PK_MARIO/PK_MARIO/Character.h
@@ -15,7 +15,7 @@ protected:
 	//virtual bool fire()=0;
 
 public:
-	Character(int x1, int y1, bool exist1, TYPES::FieldType type1, TYPES::DisplayLayer layer1,int hp1,double speed1) :
+	Character(float x1, float y1, bool exist1, TYPES::FieldType type1, TYPES::DisplayLayer layer1,int hp1,double speed1) :
 		Field::Field(x1,y1,exist1, type1, layer1),hp(hp1),speed(speed1)
 	{
 	};
